fix: connection renamed to connections
